Implement execute-javascript-with-promise-callback.
Fix newlines in Javascript strings.

Without the fix, the following behavior was observed:

;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
;; LISP
;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
(iolib:with-open-socket
        (s :address-family :local
           :remote-filename (uiop:native-namestring (electron-socket-path interface)))
      (write-line message s)
      (finish-output s)
      (read-line s))

;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
;; VALUE OF MESSAGE
;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
"\\\\n"

;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
;; JAVASCRIPT
;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
const server = net.createServer((socket) => {
    socket.on('data', (data) => {
        process.stdout.write("\n");
        process.stdout.write(data.toString());
    });
});

;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
;; OUTPUT OF JAVASCRIPT
;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
"\\n"

Why was this behavior observed? Because of the Lisp reader! It was
consuming some of the backslashes. Try `(write-string "\\\\n")` and
you'll see "\\n" as the return value.

--------------

Print until newline so read-line is invoked.

The following sequence of events is observed:

LISP send-message -> JAVASCRIPT client.write -> LISP read-line

We need to write "hello world\\n" in LISP send-message such that when
we get our code back in LISP read-line we'll get "hello world\n". This
is because the Lisp reader will swallow the first backslash, thus we must
include two. This will successfully trigger read-line.

Other times we want to say something like:

"this\
is a javascript\
multiline string"

Note the backslahes. In order to encode this in Lisp we have to write

"this\\\\nis a javscript"\\\\nmultiline string". Why is that?

"\\" becomes "\" and then we need a newline which is "\n" so we are
actually sending "this\\nis a javascript \\nmultiline string" after
the Lisp reader has mutated the string. Why do we need this? Because
Parenscript incorrectly encodes multiline strings without considering
the Lisp reader. Send-message will use write-line which will not work
if we simply passed in "this\\nis a javascript\\nmultiline string".