server: Create ProtocolSocket class.
This class is used for handling message chunking within
sockets. Typically messages are chunked by 2^16. Because we may be
processing very large messages, we handle the chunking ourselves.

We pass in a socket that is receiving the data, and then a function
that is called when all the data is received. The data end is delineated
by an end of transmission ascii control character.