Create SynchronousSocket class.
Create a class that uses execFileSync to synchronously read
information from a socket.

execFileSync calls query.js with some data. query.js will write this
data to the socket and read the response. query.js will then spit out
this data to stdout, from where it can be read by
execFileSync. Inspired by the Node.js sync-socket library.

server.js: Optionally specify a timeout.

Allow dynamic definition of the query.js script path.

This is useful for when server.js is moved to a tmp file for the
creation of protocols. The SynchronousSocket class must be able to
have a path to the module to execute it.