browser-view/window: Allow register-before-input-event bubbling control.
With this new interface, register-before-input-event allows the Lisp
side of cl-electron to control whether event.preventDefault() will be
invoked, and thus whether an event WILL/WILL NOT propagate throughout
the renderer. For example, if we want to intercept C-a in Lisp, we can
return "t" from our callback in register-before-input-event, and we'll
be able to avoid C-a being handled by the renderer.